backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=5.1.0",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "bcrypt>=4.0.0,<5.0.0",
    "cloudinary>=1.44.1",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "jose>=1.0.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
